DistinctUntilChanged: Add null support and test fixes
Changed test name styling to that of PR#25 and improve the keySelector test
